refactor(web): give the peeking eyes one implementation instead of two
`AssistantPeekingEyes` was a near-verbatim copy of `OnboardingPeekingEyes`:
the same sizing math, blink state machine, cursor parallax, entrance motion
and SVG render, with the same constants written twice (the entrance hand-off
appeared as `STAGE_CENTER_FRACTION = 0.4` in one and `PICKER_CENTER_VH = 40`
in the other). `create-personality-step.tsx` restated the derived visible
fraction a third time as a literal `0.3 * (1 - 0.25)`, with a comment asking
the reader to keep it in sync by hand.

Extract the presentation into `components/avatar/peeking-eyes.tsx`, which
takes the eye art and the stage box as props. Both components become thin
adapters that resolve the art from their own source, and both delete their
copy. `EYES_VISIBLE_FRACTION` now has one owner that both reserve callers
import.

Also drops `settleBlink`: every shipped call site passed it equal to
`entrance`, so the component derives it.

Observable behavior is unchanged. Verified by measuring the emitted DOM
before and after: the onboarding Resting and InsetByASafeArea stories and a
side-by-side render of the old and new assistant components all produce
identical element trees, class strings, style strings and geometry.
